+++ openquake/calculators/tests/classical_risk_test.py
@@ -23,7 +23,7 @@ from openquake.qa_tests_data.classical_risk import (
 from openquake.baselib.general import writetmp
 from openquake.calculators.tests import CalculatorTestCase
 from openquake.commonlib.writers import scientificformat
-from openquake.commonlib.export import export
+from openquake.calculators.export import export
 from openquake.calculators.views import view
