+++ openquake/calculators/tests/scenario_risk_test.py
@@ -25,7 +25,7 @@ from openquake.qa_tests_data.scenario_risk import (
 from openquake.baselib.general import writetmp
 from openquake.calculators.tests import CalculatorTestCase
 from openquake.calculators.views import view
-from openquake.commonlib.export import export
+from openquake.calculators.export import export
 
 
 class ScenarioRiskTestCase(CalculatorTestCase):
