+++ openquake/server/views.py
@@ -46,7 +46,7 @@ from openquake.baselib.general import groupby, writetmp
 from openquake.baselib.python3compat import unicode
 from openquake.commonlib import nrml, readinput, oqvalidation
 from openquake.baselib.parallel import TaskManager, safely_call
-from openquake.commonlib.export import export
+from openquake.calculators.export import export
 from openquake.engine import __version__ as oqversion
 from openquake.engine.export import core
 from openquake.engine import engine, logs
